src/requirements.txt
Flask==3.0.0
pymongo==4.6.1
python-dotenv==0.19.0
PyJWT==2.8.0 
arrow==1.2.0
